% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_metadata.R
\encoding{UTF-8}
\name{calculate_equivalents}
\alias{calculate_equivalents}
\title{Calculate Emission Equivalents}
\usage{
calculate_equivalents(emissions_kg)
}
\arguments{
\item{emissions_kg}{Emissions in kilograms of CO2.}
}
\value{
A list with `lightbulb_seconds`, `phone_charge_seconds`, and `car_km`.
}
\description{
Converts emissions (in kg CO2) into human-readable equivalents
}
\details{
Calculate Emission Equivalents
}
\keyword{internal}
