% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionsTracker.R
\name{EmissionsTracker}
\alias{EmissionsTracker}
\title{EmissionsTracker R6 Class}
\value{
An R6 object of class `EmissionsTracker`.
}
\description{
The `EmissionsTracker` class allows you to estimate CO2 emissions
based on the duration of a computation and a fixed energy factor.
}
\details{
A lightweight tracker for estimating carbon emissions (kg CO2) of a code block
by measuring the elapsed time and applying a fixed energy-to-emission factor.
}
\examples{
tracker <- EmissionsTracker$new()
tracker$start()
Sys.sleep(1)
tracker$stop()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{start_time}}{POSIX time when tracking began.}

\item{\code{end_time}}{POSIX time when tracking stopped.}

\item{\code{emissions}}{Estimated emissions in kg CO2.}

\item{\code{duration}}{Duration of the computation (in seconds).}

\item{\code{energy_factor}}{Emissions per second (default is 0.0002 kg/sec).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EmissionsTracker-new}{\code{EmissionsTracker$new()}}
\item \href{#method-EmissionsTracker-start}{\code{EmissionsTracker$start()}}
\item \href{#method-EmissionsTracker-stop}{\code{EmissionsTracker$stop()}}
\item \href{#method-EmissionsTracker-clone}{\code{EmissionsTracker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmissionsTracker-new"></a>}}
\if{latex}{\out{\hypertarget{method-EmissionsTracker-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmissionsTracker$new(energy_factor = 2e-04)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{energy_factor}}{A numeric value representing emissions per second (kg CO2/sec).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmissionsTracker-start"></a>}}
\if{latex}{\out{\hypertarget{method-EmissionsTracker-start}{}}}
\subsection{Method \code{start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmissionsTracker$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmissionsTracker-stop"></a>}}
\if{latex}{\out{\hypertarget{method-EmissionsTracker-stop}{}}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmissionsTracker$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmissionsTracker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EmissionsTracker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmissionsTracker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
