% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_metadata.R
\name{get_country_full_name}
\alias{get_country_full_name}
\title{Get Full Country Name from ISO Code}
\usage{
get_country_full_name(iso_code)
}
\arguments{
\item{iso_code}{A 2-letter country ISO code (e.g., "US", "DE").}
}
\value{
A full country name as a string.
}
\description{
Converts a two-letter ISO country code (e.g., "DE") to the full country name
(e.g., "Germany") using the `countrycode` package.
}
\keyword{internal}
