% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_metadata.R
\name{get_geo_info}
\alias{get_geo_info}
\title{Get Geolocation Metadata from IP}
\usage{
get_geo_info()
}
\value{
A named list with `country_iso_code`, `country_name`, `region`, `city`, `latitude`, and `longitude`.
}
\description{
Queries the `ipinfo.io` API to retrieve location-based metadata (country, region, city, coordinates).
}
\keyword{internal}
